+++ app/src/main/java/ca/pkay/rcloneexplorer/AppCompatPreferenceActivity.java
@@ -1,7 +1,10 @@
 package ca.pkay.rcloneexplorer;
 
+import android.app.ActivityManager;
 import android.content.SharedPreferences;
 import android.content.res.Configuration;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.os.Bundle;
 import android.preference.PreferenceActivity;
 import android.preference.PreferenceManager;
@@ -31,6 +34,10 @@ public abstract class AppCompatPreferenceActivity extends PreferenceActivity {
         int customAccentColor = sharedPreferences.getInt("pref_key_color_accent", -1);
         getTheme().applyStyle(CustomColorHelper.getPrimaryColorTheme(this, customPrimaryColor), true);
         getTheme().applyStyle(CustomColorHelper.getAccentColorTheme(this, customAccentColor), true);
+        // set recents app color to the primary color
+        Bitmap bm = BitmapFactory.decodeResource(getResources(), R.mipmap.ic_launcher_round);
+        ActivityManager.TaskDescription taskDesc = new ActivityManager.TaskDescription(getString(R.string.app_name), bm, customPrimaryColor);
+        setTaskDescription(taskDesc);
         super.onCreate(savedInstanceState);
     }