+++ app/src/main/java/ca/pkay/rcloneexplorer/ChangelogActivity.java
@@ -1,6 +1,9 @@
 package ca.pkay.rcloneexplorer;
 
+import android.app.ActivityManager;
 import android.content.SharedPreferences;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.v7.app.ActionBar;
@@ -35,6 +38,11 @@ public class ChangelogActivity extends AppCompatActivity {
         int customAccentColor = sharedPreferences.getInt("pref_key_color_accent", -1);
         getTheme().applyStyle(CustomColorHelper.getPrimaryColorTheme(this, customPrimaryColor), true);
         getTheme().applyStyle(CustomColorHelper.getAccentColorTheme(this, customAccentColor), true);
+
+        // set recents app color to the primary color
+        Bitmap bm = BitmapFactory.decodeResource(getResources(), R.mipmap.ic_launcher_round);
+        ActivityManager.TaskDescription taskDesc = new ActivityManager.TaskDescription(getString(R.string.app_name), bm, customPrimaryColor);
+        setTaskDescription(taskDesc);
     }
 
     @Override